Refactor quiz data structure and constants
- Moved the QUESTIONS_PER_GAME constant to a new file for better maintainability.
- Updated all quiz data files to remove individual QUESTIONS_PER_GAME exports and use the new DEFAULT_QUESTIONS_PER_GAME constant.
- Cleaned up imports in quizGameConfigs.ts to reflect the changes in the data files.
- Ensured consistent usage of the new constant across all quiz types including spelling, sports, languages, and others.
