+++ gamesformykids/app/games/taki/takiDeck.ts
@@ -1,4 +1,5 @@
 import type { TakiCard, CardColor, CardValue } from './takiTypes';
+import { shuffle } from '@/lib/utils';
 
 export const CARD_COLORS: CardColor[] = ['red', 'green', 'blue', 'yellow'];
 
@@ -18,13 +19,4 @@ export function buildDeck(): TakiCard[] {
     cards.push({ id: id(), color: 'wild', value: 'king' });
   }
   return cards;
-}
-
-export function shuffle<T>(arr: T[]): T[] {
-  const a = [...arr];
-  for (let i = a.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [a[i], a[j]] = [a[j], a[i]];
-  }
-  return a;
 }
