+++ gamesformykids/components/game/quiz/games/israel/useIsraelGame.ts
@@ -1,7 +1,8 @@
 'use client';
 
 import { createCategoryIndexQuizHook } from '@/lib/quiz/createCategoryIndexQuizHook';
-import { ISRAEL_QUESTIONS, QUESTIONS_PER_GAME, CATEGORIES } from './data/questions';
+import { ISRAEL_QUESTIONS, CATEGORIES } from './data/questions';
+import { QUESTIONS_PER_GAME } from '@/lib/quiz/constants';
 
 export const useIsraelGame = createCategoryIndexQuizHook({
   questions: ISRAEL_QUESTIONS,
