+++ gamesformykids/lib/quiz/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_QUESTIONS_PER_GAME = 10;
+export const QUESTIONS_PER_GAME = DEFAULT_QUESTIONS_PER_GAME;
