+++ gamesformykids/lib/quiz/quizGameConfigs.tsx
@@ -3,22 +3,23 @@ import { shuffle } from '@/lib/utils';
 import type { QuizTheme } from '@/components/game/quiz/quizTheme';
 import type { GameType } from '@/lib/types/core/base';
 import FractionBar from '@/components/game/quiz/FractionBar';
+import { DEFAULT_QUESTIONS_PER_GAME } from '@/lib/quiz/constants';
 
 // ── Data imports ──────────────────────────────────────────────────────────────
-import { RIDDLES, QUESTIONS_PER_GAME as RPG } from '@/lib/quiz/data/riddles';
-import { CAPITAL_QUESTIONS as CAPITALS, QUESTIONS_PER_GAME as CPG } from '@/lib/quiz/data/capitals';
-import { SPELLING_WORDS, QUESTIONS_PER_GAME as SPGPG } from '@/lib/quiz/data/spelling';
-import { FRACTION_QUESTIONS, QUESTIONS_PER_GAME as FPGPG } from '@/lib/quiz/data/fractions';
-import { EMOTION_QUESTIONS, QUESTIONS_PER_GAME as EPGPG } from '@/lib/quiz/data/emotions';
-import { INSTRUMENTS, QUESTIONS_PER_GAME as IPGPG } from '@/lib/quiz/data/instruments';
-import { LANGUAGE_QUESTIONS, QUESTIONS_PER_GAME as LPGPG } from '@/lib/quiz/data/world-languages';
-import { OPPOSITE_WORDS, QUESTIONS_PER_GAME as OPGPG } from '@/lib/quiz/data/opposites';
-import { SPORTS_QUESTIONS, QUESTIONS_PER_GAME as SQPGPG } from '@/lib/quiz/data/sports-quiz';
-import { CONTINENT_QUESTIONS, QUESTIONS_PER_GAME as CQPGPG } from '@/lib/quiz/data/continents';
-import { NUTRITION_QUESTIONS, QUESTIONS_PER_GAME as NQPGPG } from '@/lib/quiz/data/healthy-food';
-import { FAMILY_QUESTIONS, QUESTIONS_PER_GAME as FQPGPG } from '@/lib/quiz/data/family';
-import { ENGLISH_WORDS, QUESTIONS_PER_GAME as EWPGPG } from '@/lib/quiz/data/english-words';
-import { QUIZ_QUESTIONS, SHAPES_3D, QUESTIONS_PER_GAME as S3PGPG } from '@/lib/quiz/data/shapes-3d';
+import { RIDDLES } from '@/lib/quiz/data/riddles';
+import { CAPITAL_QUESTIONS as CAPITALS } from '@/lib/quiz/data/capitals';
+import { SPELLING_WORDS } from '@/lib/quiz/data/spelling';
+import { FRACTION_QUESTIONS } from '@/lib/quiz/data/fractions';
+import { EMOTION_QUESTIONS } from '@/lib/quiz/data/emotions';
+import { INSTRUMENTS } from '@/lib/quiz/data/instruments';
+import { LANGUAGE_QUESTIONS } from '@/lib/quiz/data/world-languages';
+import { OPPOSITE_WORDS } from '@/lib/quiz/data/opposites';
+import { SPORTS_QUESTIONS } from '@/lib/quiz/data/sports-quiz';
+import { CONTINENT_QUESTIONS } from '@/lib/quiz/data/continents';
+import { NUTRITION_QUESTIONS } from '@/lib/quiz/data/healthy-food';
+import { FAMILY_QUESTIONS } from '@/lib/quiz/data/family';
+import { ENGLISH_WORDS } from '@/lib/quiz/data/english-words';
+import { QUIZ_QUESTIONS, SHAPES_3D } from '@/lib/quiz/data/shapes-3d';
 
 // ─────────────────────────────────────────────────────────────────────────────
 
@@ -62,7 +63,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         <p className="text-purple-600 font-bold text-sm mt-1">שעון ⏰</p>
       </div>
     ),
-    questions: RIDDLES, questionsPerGame: RPG,
+    questions: RIDDLES, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.answer, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.answer,
     getCorrectLabel: (q) => q.answer,
@@ -82,7 +83,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         ))}
       </div>
     ),
-    questions: CAPITALS, questionsPerGame: CPG,
+    questions: CAPITALS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.capital, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.capital,
     getCorrectLabel: (q) => q.capital,
@@ -97,7 +98,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
     gameType: 'spelling', emoji: '📝', title: 'כתיב עברי',
     description: 'בחר את האיות הנכון לכל מילה!', theme: 'rose',
     buttonLabel: '✏️ בואו נכתוב!',
-    questions: SPELLING_WORDS, questionsPerGame: SPGPG,
+    questions: SPELLING_WORDS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.word, ...q.wrong]),
     isCorrect: (c, q) => c === q.word,
     getCorrectLabel: (q) => q.word,
@@ -120,7 +121,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         <p className="text-purple-800 font-bold mt-1">חצי (1/2) ✓</p>
       </div>
     ),
-    questions: FRACTION_QUESTIONS, questionsPerGame: FPGPG,
+    questions: FRACTION_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.display, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.display,
     getCorrectLabel: (q) => q.display,
@@ -142,7 +143,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         ))}
       </div>
     ),
-    questions: EMOTION_QUESTIONS, questionsPerGame: EPGPG,
+    questions: EMOTION_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.emotion, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.emotion,
     getCorrectLabel: (q) => q.emotion,
@@ -163,7 +164,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         ))}
       </div>
     ),
-    questions: INSTRUMENTS, questionsPerGame: IPGPG,
+    questions: INSTRUMENTS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.instrument, ...q.wrongInstruments]),
     isCorrect: (c, q) => c === q.instrument,
     getCorrectLabel: (q) => q.instrument,
@@ -189,7 +190,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
       </div>
     ),
     buttonLabel: '🌍 בואו נלמד!',
-    questions: LANGUAGE_QUESTIONS, questionsPerGame: LPGPG,
+    questions: LANGUAGE_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.language, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.language,
     getCorrectLabel: (q) => q.language,
@@ -210,7 +211,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
       </div>
     ),
     buttonLabel: '🙃 משחק ניגודים!',
-    questions: OPPOSITE_WORDS, questionsPerGame: OPGPG,
+    questions: OPPOSITE_WORDS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.opposite, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.opposite,
     getCorrectLabel: (q) => q.opposite,
@@ -231,7 +232,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         ))}
       </div>
     ),
-    questions: SPORTS_QUESTIONS, questionsPerGame: SQPGPG,
+    questions: SPORTS_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => [...q.answers],
     isCorrect: (c, q) => c === q.answers[q.correctIndex],
     getCorrectLabel: (q) => q.answers[q.correctIndex],
@@ -245,7 +246,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
   continents: {
     gameType: 'continents', emoji: '🌍', title: 'יבשות העולם',
     description: 'ענה על שאלות על יבשות כדור הארץ!', theme: 'teal',
-    questions: CONTINENT_QUESTIONS, questionsPerGame: CQPGPG,
+    questions: CONTINENT_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => [...q.answers],
     isCorrect: (c, q) => c === q.answers[q.correctIndex],
     getCorrectLabel: (q) => q.answers[q.correctIndex],
@@ -258,7 +259,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
   'healthy-food': {
     gameType: 'healthy-food', emoji: '🥗', title: 'אוכל בריא',
     description: 'למד על תזונה בריאה!', theme: 'green',
-    questions: NUTRITION_QUESTIONS, questionsPerGame: NQPGPG,
+    questions: NUTRITION_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => [...q.answers],
     isCorrect: (c, q) => c === q.answers[q.correctIndex],
     getCorrectLabel: (q) => q.answers[q.correctIndex],
@@ -271,7 +272,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
   family: {
     gameType: 'family', emoji: '👨‍👩‍👧‍👦', title: 'המשפחה',
     description: 'ענה על שאלות על קשרי משפחה!', theme: 'rose',
-    questions: FAMILY_QUESTIONS, questionsPerGame: FQPGPG,
+    questions: FAMILY_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => [...q.answers],
     isCorrect: (c, q) => c === q.answers[q.correctIndex],
     getCorrectLabel: (q) => q.answers[q.correctIndex],
@@ -291,7 +292,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
         ))}
       </div>
     ),
-    questions: ENGLISH_WORDS, questionsPerGame: EWPGPG,
+    questions: ENGLISH_WORDS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => shuffle([q.english, ...q.wrongOptions]),
     isCorrect: (c, q) => c === q.english,
     getCorrectLabel: (q) => q.english,
@@ -305,7 +306,7 @@ export const QUIZ_GAME_CONFIGS: Partial<Record<GameType, QuizGameConfig<any>>> =
   'shapes-3d': {
     gameType: 'shapes-3d', emoji: '📦', title: 'גופים גיאומטריים',
     description: 'זהה גופים תלת-ממדיים!', theme: 'violet',
-    questions: QUIZ_QUESTIONS, questionsPerGame: S3PGPG,
+    questions: QUIZ_QUESTIONS, questionsPerGame: DEFAULT_QUESTIONS_PER_GAME,
     getChoices: (q) => {
       const shape = SHAPES_3D.find(s => s.id === q.shapeId)!;
       return shuffle([q.answer, ...shape.wrongOptions]);
